PortfolioProjects/VibeCoding/cv-matcher/requirements.txt
streamlit
sentence-transformers
scikit-learn
PyPDF2
python-docx
pytest
pandas
flask
